Correct logging that contained non printable chars
While debugging we find out in logs that instead of the number of the
pods running there was a strange char shown. We found that chars printed
where ^C and ^B, which respectively corresponds to the char 3 and 2 of
ascii table. With this fix we use correct fmt function to print out the
correct value.