[#845] Make: check for go version compatibility
Building with golang 1.21 requires changes to the go.mod file. This is
most likely due to a change in behavior in golang's gomod (see
https://go.dev/doc/modules/gomod-ref).

Since the migration to the newer version of go hasn't yet been made, and
to help any user who might have a > 1.20 version of go on their system,
the Makefile now brings in a new check to verify that the go version
matches the one of the go.mod.